requirements.txt
anyjson>=0.3.3
argparse
Babel>=1.3
boto>=2.32.1
eventlet>=0.16.1,!=0.17.0
greenlet>=0.3.2
httplib2>=0.7.5
iso8601>=0.1.9
jsonschema>=2.0.0,<3.0.0
lxml>=2.3
oslo-config>=1.9.3,<1.10.0  # Apache-2.0
oslo-concurrency>=1.8.0,<1.9.0         # Apache-2.0
oslo-db>=1.7.0,<1.8.0  # Apache-2.0
oslo-log>=1.0.0,<1.1.0  # Apache-2.0
oslo-messaging>=1.8.0,<1.9.0  # Apache-2.0
oslo-serialization>=1.4.0,<1.5.0               # Apache-2.0
oslo-utils>=1.4.0,<1.5.0                       # Apache-2.0
paramiko>=1.13.0
Paste
PasteDeploy>=1.5.0
pbr>=0.6,!=0.7,<1.0
pyasn1
pyasn1_modules
python-cinderclient>=1.1.0,<1.2.0
python-glanceclient>=0.15.0,<0.18.0
python-keystoneclient>=1.2.0,<1.4.0
python-neutronclient>=2.3.11,<2.5.0
python-novaclient>=2.22.0,<2.24.0
Routes>=1.12.3,!=2.0
six>=1.9.0
SQLAlchemy>=0.9.7,<=0.9.99
sqlalchemy-migrate>=0.9.5
stevedore>=1.3.0,<1.4.0  # Apache-2.0
WebOb>=1.2.3
metrics>=0.1
